chore(achievements): Extract achievements page logic
